Pass compute_error and epsilon as arguments
This PR adds a `compute_error` field to `TruncatedAlgorithm` and
changes `eig_trunc!`, `eigh_trunc!`, and `svd_trunc!` (and their
non-mutating counterparts) to accept an array `epsilon` as part
of their arguments. The purpose of this is twofold: to make handling
the truncation error easier in AD, and to avoid forcing GPU
synchronization.